aprcl

+++ src/aprcl/unity_zp_coeff.c
@@ -55,7 +55,7 @@ unity_zp_coeff_add_ui(unity_zp f, ulong ind, ulong x)
 void
 unity_zp_coeff_inc(unity_zp f, ulong ind)
 {
-    if (ind >= f->poly->length)
+    if (ind >= (ulong) f->poly->length)
     {
         fmpz_mod_poly_set_coeff_ui(f->poly, ind, 1, f->ctx);
         return;
@@ -69,7 +69,7 @@ unity_zp_coeff_inc(unity_zp f, ulong ind)
 void
 unity_zp_coeff_dec(unity_zp f, ulong ind)
 {
-    if (ind >= f->poly->length)
+    if (ind >= (ulong) f->poly->length)
     {
         fmpz_mod_poly_set_coeff_si(f->poly, ind, -1, f->ctx);
         return;
